Partial rollback
Since Folder is a foreign key, using a value of 0 violates the constraint
(no Folder with ID will exist). Workarounds would be either creating a new
column to enforce uniqueness with in a multi-column index (similar to how
we disable unique constraints when discardable objects are discarded), revert
to not using a FK column, or (if possible) creating a Folder with ID of 0
in each tenant, probably when the first folder is created.