Form schema now a text column
Since JSON columns don't preserve the raw JSON, we lose things like order
of object keys. It makes a bit more sense to store the raw JSON since we
don't need to index anything or make use of anything that would require
it be a JSON column in the database. Also don't want to pollute the
JSON schema with extra data to specify order.