Switch UpdateDocumentSchema to use JSON schema validation
SmartParams is great but I keep running into issues. The schema of request
bodies can vary a lot and since it uses Dry::Types (which doesn't have
built-in support for null values), I've had to implement some additional
behavior in the gem that continues to present issues. Until I can put in more
time to diagnose and correct the issues, I'll revert to a more reliable
solution by using JSON schemas to validate the bodies of various requests.
It's a less DRY, more tedious, but will be reliable and provide the same
benefit as SmartParams.